Rename getFirstIpAddressFromHeader method to getIpAddressFromHeader
As the leftmost IPs on the Forwarded header list can be spoofed, we are now taking the rightmost IPs from the list.

The library now includes a fifth parameter called $trustedProxiesCount, which is used to ignore a certain number of IPs from the Forwarded header list, so trusted or known IPs can be ignored.